Remove TopBar, reuse StaticAppBar #2373
