+++ src/components/NavigationBar/Styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+import IconButton from '@material-ui/core/IconButton';
+
+export const StyledIconButton = styled(IconButton)`
+  border-radius: 2px;
+  padding: 0.5rem;
+`;
